+++ bin/dx_get_appliance.pl
@@ -84,7 +84,9 @@ if (defined($details)) {
     {'VDBs#',      8},
     {'Total Objects', 8},
     {'vCpu',  8},
-    {'vMem [GB]',  9}
+    {'vMem [GB]',  9},
+    {'UUID',40},
+    {'Type',20}
   );
 } else {
   $output->addHeader(
@@ -113,6 +115,8 @@ for my $engine ( sort (@{$engine_list}) ) {
       $output->addLine(
         $engine,
         $status,
+        "",
+        "",
         "",
         "",
         "",
@@ -166,7 +170,9 @@ for my $engine ( sort (@{$engine_list}) ) {
       scalar(@vdb),
       scalar(@dsource) +  scalar(@vdb),
       $system->getvCPU(),
-      sprintf("%8.2f",$system->getvMem())
+      sprintf("%8.2f",$system->getvMem()),
+      $system->getUUID(),
+      $system->getEngineType()
     );
   } else {
     $output->addLine(
@@ -288,10 +294,10 @@ Display a Delphix Engine summary
 
 Display a Delphix Engine details 
 
-  dx_get_appliance.pl -d dxt1 -details
+ dx_get_appliance.pl -d dxt1 -details
 
-  Appliance            Status   Version  Total (GB) Used (GB)  Free (GB)  PctUsed(%) dSource# VDBs#    Total Ob vCpu     vMem [GB]
-  -------------------- -------- -------- ---------- ---------- ---------- ---------- -------- -------- -------- -------- ---------
-  dxt1                 UP       6.0.11.0 33.42      9.17       24.25      27.44      1        2        3        4           15.88
+ Appliance            Status   Version  Total (GB) Used (GB)  Free (GB)  PctUsed(%) dSource# VDBs#    Total Ob vCpu     vMem [GB] UUID
+ -------------------- -------- -------- ---------- ---------- ---------- ---------- -------- -------- -------- -------- --------- ----------------------------------------
+ dxt1                 UP       6.0.12.1 21.80      7.62       14.18      34.96      1        0        1        2            8.00  564d754d-eb0f-bb3e-15a2-f45c08d0ae24
 
 =cut